+++ CHANGELOG.md
@@ -1,3 +1,14 @@
-# `0.1.1`
+# Change Log
+
+## `0.2.0` - options verification
+
+- As scp user, I want to default port=`22` if not provided.
+- As scp user, I want to default localpath=current directory (`.`) if not provided.
+- As scp user, I want to default remotepath=current directory (`.`) if not provided.
+- As scp user, I want to have error if username not provided.
+- As scp user, I want to have error if password not provided.
+- As scp user, I want to have error if hostname not provided.
+
+## `0.1.1`
 
 - As scp user, I want to download all files from a remote directory to a local directory.